Introduce gammapy.utils.docs and fix GammapyExtraImageDirective
